% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3method-as.matrix.R
\name{as.matrix.viterbi}
\alias{as.matrix.viterbi}
\title{Matrix coercion for qtl2 objects}
\usage{
\method{as.matrix}{viterbi}(x)
}
\arguments{
\item{x}{The object to coerce.}
}
\description{
Convert the output of \code{qtl2} functions to \code{matrix}.
}
