% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_markers.R
\name{get_markers}
\alias{get_markers}
\alias{get_markers.cross2}
\title{Extract markers from cross2 object}
\usage{
get_markers(x)

\method{get_markers}{cross2}(x)
}
\arguments{
\item{x}{An object of class "cross2".}
}
\value{
a tibble
}
\description{
This function gets both genetic and physical position of all markers in a \code{cross2}
object.
}
