% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{qtl2helper_utils}
\alias{qtl2helper_utils}
\alias{.join_map_by_marker}
\title{Utility functions for qtl2helper}
\usage{
.join_map_by_marker(.tbl, .map)
}
\arguments{
\item{.tbl}{a \code{data.frame} or \code{tibble} to add the map data to.
Should contain a column named "marker".}

\item{.map}{the marker map, for example from \code{qtl2::insert_pseudomarkers()}.}
}
\description{
These are internal utility functions (not exported).

\code{.join_map_by_marker()} joins a qtl2 marker map to a table.
}
