% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{tidy.scan1}
\alias{tidy.scan1}
\alias{tidy.scan1coef}
\alias{tidy.calc_genoprob}
\alias{tidy.scan1perm}
\title{Tidy R/qtl2 objects}
\usage{
\method{tidy}{scan1}(x, map = NULL)

\method{tidy}{scan1coef}(x, map = NULL)

\method{tidy}{calc_genoprob}(x, map = NULL)

\method{tidy}{scan1perm}(x, alpha = 0.05)
}
\arguments{
\item{x}{The object to coerce.}

\item{map}{An optional list of vectors of marker positions, as produced by
qtl2::insert_pseudomarkers() or the \verb{$gmap} element of a \code{cross2} object.}

\item{alpha}{Vector of significance levels (for \code{scan1perm} objects).}
}
\description{
Convert the output of \code{qtl2} functions to a \code{tibble} in \emph{long} format.
These are generally useful for downstream plotting analysis.
}
\details{
At the moment the output from the following \code{R/qtl2} functions are recognised:
\itemize{
\item \code{scan1()}
\item \code{scan1coef()}
}
}
