% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.join_map_by_marker}
\alias{.join_map_by_marker}
\title{Utility functions for qtl2helper}
\usage{
.join_map_by_marker(.tbl, .map)
}
\arguments{
\item{.tbl}{a `data.frame` or `tibble` to add the map data to.
Should contain a column named "marker".}

\item{.map}{the marker map, for example from `qtl2::insert_pseudomarkers()`.}
}
\description{
These functions are not exported.
}
